// src/app/models/kytarsiv.ts
export class RecordArchive {
    kyt_id?: number;
    kyt_table?: string;
    kyt_kid?: number;
    kyt_kid2?: number;
    kyt_ksira?: number;
    kyt_ksira2?: number;
    kyt_user?: string;
    kyt_col_name?: string;
    kyt_col_adi?: string;
    kyt_time?: string;
    kyt_islem?: number;
    kyt_kod?: string;
    kyt_clas_name?: string;
    kyt_grs_no?: number;
    kyt_grs_win?: string;
    kyt_grs_tur?: number;
    kyt_grs_tab?: number;
    kyt_grs_tabp?: number;
    kyt_dw_name?: string;
    kyt_ilk_veri?: string;
    kyt_son_veri?: string;
    kyt_srk_no?: number;
  }
  